updated agent code structure
