+++ agent/Dockerfile
@@ -11,11 +11,11 @@ RUN apt-get update && apt-get install -y build-essential curl && rm -rf /var/lib
 COPY requirements.txt .
 RUN pip install --no-cache-dir -r requirements.txt
 
-# Copy the FastAPI app code
-COPY . .
+# Copy the whole app directory
+COPY . ./agent
 
 # Expose FastAPI port
 EXPOSE 8000
 
-# Start FastAPI with Uvicorn
-CMD ["uvicorn", "agent:app", "--host", "0.0.0.0", "--port", "8000"]
+# Start FastAPI with Uvicorn using module path
+CMD ["uvicorn", "agent.main:app", "--host", "0.0.0.0", "--port", "8000"]
