+++ agent/models/request.py
@@ -1,0 +1,4 @@
+from pydantic import BaseModel
+
+class PromptRequest(BaseModel):
+    prompt: str
